backend/requirements.txt
flask
supabase
psycopg2
python-dotenv
google-genai
PyPDF2
langchain==0.1.0
langchain-openai==0.0.2
langchain-core
openai
nltk
langgraph
requests
langchain-google-genai
flask-cors
torch
torchvision
torchaudio
transformers
pillow
pdf2image
